refactor(renderer): split App.vue responsibilities into feature composables
App.vue から各 feature 固有の購読・watcher を抜き、bootstrap composable の呼び出しだけにする。

- features/worktree/useGitStatusSync: dir 切替と Claude state 遷移で loadGitStatus
- features/filer/useFsWatchSync: rpcFsWatch/Unwatch 付替
- features/sidebar/useGozdOpenHandler: gozdOpen 購読 + addRepo
- features/sidebar/useRepoContextKey: selectedIsGitRepo → context key
- features/layout/useNotifySubscription: notify push → toast
- features/layout/useCommandErrorBridge: command error → toast
- useGitStatusStore.loadGitStatus に gen counter を追加し、並行呼び出しで stale なレスポンスが新しい結果を上書きする race を抑止
