+++ apps/renderer/src/features/filer/index.ts
@@ -1,4 +1,5 @@
 export { default as FilerPane } from "./FilerPane.vue";
-export { rpcFsReadFile, rpcFsReadFileAbsolute, rpcFsUnwatch, rpcFsWatch } from "./rpc";
+export { rpcFsReadFile, rpcFsReadFileAbsolute,   } from "./rpc";
 export type { FsChangePayload } from "./rpc";
 export { getFileIconName, getIconUrl } from "./useFileIcon";
+export { useFsWatchSync } from "./useFsWatchSync";
