+++ apps/renderer/src/features/layout/index.ts
@@ -1,3 +1,5 @@
 export { default as MainLayout } from "./MainLayout.vue";
 export { default as ResizeHandle } from "./ResizeHandle.vue";
 export { rpcPickAndOpen } from "./rpc";
+export { useCommandErrorBridge } from "./useCommandErrorBridge";
+export { useNotifySubscription } from "./useNotifySubscription";
