+++ apps/renderer/src/features/sidebar/index.ts
@@ -1,3 +1,5 @@
 export { default as SidebarPane } from "./SidebarPane.vue";
-export { rpcCreateWorktree, rpcGitBranchList, rpcGitWorktreeList, rpcTaskAdd } from "./rpc";
+export { rpcCreateWorktree,  rpcGitWorktreeList, rpcTaskAdd } from "./rpc";
 export type { BranchChangePayload } from "./rpc";
+export { useGozdOpenHandler } from "./useGozdOpenHandler";
+export { useRepoContextKey } from "./useRepoContextKey";
