+++ apps/renderer/src/features/worktree/index.ts
@@ -4,6 +4,7 @@ export type { GitStatusChangePayload } from "./rpc";
 export { default as StatusIcons } from "./StatusIcons.vue";
 export { useWorktreeStore } from "./useWorktreeStore";
 export { useGitStatusStore } from "./useGitStatusStore";
+export { useGitStatusSync } from "./useGitStatusSync";
 export {
   computeStatusIcons,
   resolveDirectoryGitChange,
